+++ player/js/elements/BaseElement.js
@@ -14,10 +14,17 @@ BaseElement.prototype.checkMasks = function(){
     return false;
 }
 
+BaseElement.prototype.checkParenting = function(){
+    if(this.data.parent !== undefined){
+        this.comp.buildElementParenting(this, this.data.parent);
+    }
+}
+
 BaseElement.prototype.prepareFrame = function(num){
     if(this.data.ip - this.data.st <= num && this.data.op - this.data.st > num)
     {
         if(this.isVisible !== true){
+            this.elemMdf = true;
             this.globalData.mdf = true;
             this.isVisible = true;
             this.firstFrame = true;
@@ -27,6 +34,7 @@ BaseElement.prototype.prepareFrame = function(num){
         }
     }else{
         if(this.isVisible !== false){
+            this.elemMdf = true;
             this.globalData.mdf = true;
             this.isVisible = false;
         }
@@ -35,6 +43,7 @@ BaseElement.prototype.prepareFrame = function(num){
     for(i=0;i<len;i+=1){
         this.dynamicProperties[i].getValue();
         if(this.dynamicProperties[i].mdf){
+            this.elemMdf = true;
             this.globalData.mdf = true;
         }
     }
@@ -54,9 +63,49 @@ BaseElement.prototype.prepareFrame = function(num){
     }
     */
 
+
     this.currentFrameNum = num*this.data.sr;
     return this.isVisible;
 };
+
+BaseElement.prototype.globalToLocal = function(pt){
+    var transforms = [];
+    transforms.push(this.finalTransform);
+    var flag = true;
+    var comp = this.comp;
+    while(flag){
+        if(comp.finalTransform){
+            if(comp.data.hasMask){
+                transforms.splice(0,0,comp.finalTransform);
+            }
+            comp = comp.comp;
+        } else {
+            flag = false;
+        }
+    }
+    var i, len = transforms.length,ptNew;
+    for(i=0;i<len;i+=1){
+        ptNew = transforms[i].mat.applyToPointArray(0,0,0);
+        //ptNew = transforms[i].mat.applyToPointArray(pt[0],pt[1],pt[2]);
+        pt = [pt[0] - ptNew[0],pt[1] - ptNew[1],0];
+    }
+    return pt;
+};
+
+BaseElement.prototype.initExpressions = function(){
+    this.layerInterface = LayerExpressionInterface(this);
+    //layers[i].layerInterface = LayerExpressionInterface(layers[i]);
+    //layers[i].layerInterface = LayerExpressionInterface(layers[i]);
+    if(this.data.hasMask){
+        this.layerInterface.registerMaskInterface(this.maskManager);
+    }
+    if(this.data.ty === 0 || this.data.xt){
+        this.compInterface = CompExpressionInterface(this);
+    } else if(this.data.ty === 4){
+        this.layerInterface.shapeInterface = ShapeExpressionInterface.createShapeInterface(this.shapesData,this.viewData,this.layerInterface);
+    }
+}
+
 BaseElement.prototype.setBlendMode = function(){
     var blendModeValue = '';
     switch(this.data.bm){
@@ -124,21 +173,24 @@ BaseElement.prototype.init = function(){
     this.isVisible = false;
     this.currentFrameNum = -99999;
     this.lastNum = -99999;
-    this.finalTransform = {
-        mProp: PropertyFactory.getProp(this,this.data.ks,2,null,this.dynamicProperties),
-        matMdf: false,
-        opMdf: false,
-        mat: new Matrix(),
-        opacity: 1
-    };
-    this.finalTransform.op = this.finalTransform.mProp.o;
-    this.transform = this.finalTransform.mProp;
-    if(this.data.ty !== 11){
-        this.createElements();
+    if(this.data.ks){
+        this.finalTransform = {
+            mProp: PropertyFactory.getProp(this,this.data.ks,2,null,this.dynamicProperties),
+            matMdf: false,
+            opMdf: false,
+            mat: new Matrix(),
+            opacity: 1
+        };
+        this.finalTransform.op = this.finalTransform.mProp.o;
+        this.transform = this.finalTransform.mProp;
+        if(this.data.ty !== 11){
+            this.createElements();
+        }
+        if(this.data.hasMask){
+            this.addMasks(this.data);
+        }
     }
-    if(this.data.hasMask){
-        this.addMasks(this.data);
-    }
+    this.elemMdf = false;
 };
 BaseElement.prototype.getType = function(){
     return this.type;