+++ player/js/elements/canvasElements/CVImageElement.js
@@ -1,6 +1,5 @@
 function CVImageElement(data, comp,globalData){
     this.assetData = globalData.getAssetData(data.refId);
-    this.path = globalData.getPath();
     this._parent.constructor.call(this,data, comp,globalData);
     this.globalData.addPendingElement();
 }
@@ -9,6 +8,27 @@ createElement(CVBaseElement, CVImageElement);
 CVImageElement.prototype.createElements = function(){
     var imageLoaded = function(){
         this.globalData.elementLoaded();
+        if(this.assetData.w !== this.img.width || this.assetData.h !== this.img.height){
+            var canvas = document.createElement('canvas');
+            canvas.width = this.assetData.w;
+            canvas.height = this.assetData.h;
+            var ctx = canvas.getContext('2d');
+
+            var imgW = this.img.width;
+            var imgH = this.img.height;
+            var imgRel = imgW / imgH;
+            var canvasRel = this.assetData.w/this.assetData.h;
+            var widthCrop, heightCrop;
+            if(imgRel>canvasRel){
+                heightCrop = imgH;
+                widthCrop = heightCrop*canvasRel;
+            } else {
+                widthCrop = imgW;
+                heightCrop = widthCrop/canvasRel;
+            }
+            ctx.drawImage(this.img,(imgW-widthCrop)/2,(imgH-heightCrop)/2,widthCrop,heightCrop,0,0,this.assetData.w,this.assetData.h);
+            this.img = canvas;
+        }
     }.bind(this);
     var imageFailed = function(){
         this.failed = true;
@@ -18,7 +38,8 @@ CVImageElement.prototype.createElements = function(){
     this.img = new Image();
     this.img.addEventListener('load', imageLoaded, false);
     this.img.addEventListener('error', imageFailed, false);
-    this.img.src = this.path+this.assetData.p;
+    var assetPath = this.globalData.getAssetsPath(this.assetData);
+    this.img.src = assetPath;
 
     this._parent.createElements.call(this);